+++ mod_form.php
@@ -46,10 +46,20 @@ class mod_poster_mod_form extends moodleform_mod {
         $mform->addElement('header', 'general', get_string('general', 'core_form'));
 
         // Add the poster name field.
-        $mform->addElement('text', 'name', get_string('postername', 'mod_poster'), array('size' => '64'));
+        $mform->addElement('text', 'name', get_string('postername', 'mod_poster'), ['size' => '64']);
         $mform->setType('name', PARAM_TEXT);
         $mform->addRule('name', null, 'required', null, 'client');
         $mform->addRule('name', get_string('maximumchars', 'core', 255), 'maxlength', 255, 'client');
+
+        // Add the option to display the content on the course page.
+        $mform->addElement('select', 'display', get_string('display', 'mod_poster'),
+            [POSTER_DISPLAY_PAGE => get_string('displaypage', 'mod_poster'),
+                POSTER_DISPLAY_INLINE => get_string('displayinline', 'mod_poster'), ]);
+        $mform->addHelpButton('display', 'display', 'mod_poster');
+        if (!$this->courseformat->has_view_page()) {
+            $mform->setConstant('display', POSTER_DISPLAY_PAGE);
+            $mform->hardFreeze('display');
+        }
 
         // Add the show name at the view page field.
         $mform->addElement('advcheckbox', 'shownameview', get_string('shownameview', 'mod_poster'));