+++ CHANGES.md
@@ -1,3 +1,12 @@
+### 7.0.0 ###
+
+* New maintainer: bdecent gmbh.
+* Behat and PHPUnit tests updated and tested on Moodle 4.1 - 4.3.
+* Fixed minor coding style issues.
+* Added monologo for Moodle 4.x.
+* Fixed improper use of require_login method instead of require_course_login (resolves #10).
+* New feature: Display block content on course page (based on work by Harald, David und Nicklas @devcamp19 — Thanks!).
+ 
 ### 6.0.1 ###
 
 * Behat and PHPUnit tests updated and tested on Moodle 3.9 - 3.11.